Add files via upload
I’m interested in what physics tells us about the real world. There are two opposing views. One view is that our theories perfectly describe the reality we see around us — this is known as realism. But there are many cases where what we thought we knew from science turned out to be wrong, for example when Einstein’s theory of relativity trumped the Newtonian view of space and time. The opposing view is called antirealism, and says that physical theories are good at making predictions that we can use in our technologies, but they do not describe reality. I take the view that there must be something right on both sides, and that there may be a way to translate one view to another. A question I am looking at now is whether two competing theories of the structure of the universe, string theory and quantum loop gravity, have a common core.

How does your training help you think about these ideas? My Ph.D. dissertation was on the foundations of quantum mechanics. But I wanted to do something new, and I was fortunate to receive an Andrew W. Mellon Foundation award that enabled me to spend a year at the Mathematical Research Institute at the University of Utrecht in the Netherlands learning an area of mathematics known as category theory. Now I am applying these concepts to the philosophy of science and the debate between realism and antirealism.